Allow user to buy varied packages, calculate price server-side
The user now sees all available packages listed on purchase page and can
buy any quantity of them. The form directly submits data to the server,
which then calculates the price to charge based on the packages and
quantities they chose. The price is then charged to Stripe. We also save
the Stripe token.